requirements.txt
streamlit==1.56.0
psycopg2-binary==2.9.11
google-cloud-storage==3.10.0
Faker==40.13.0